pug-lexer: fix main test file
It seems <basedir> is some kind of placeholder.
For whatever reason, it's not getting substituted
and it's causing all the lexer tests to fail
because the snapshots contain the placeholder
in the filename string and the lexer is fed real paths.

Fix this by normalizing the filename given to the lexer.