filterer: remove filter aliases support
I think this was needed because the original pug syntax did not allow
some common characters in the filter names. This made it impossible to
map them to npm package names without some form of aliasing.
Now that I have introduced quoted filter syntax, this doesn't seem
to be necessary anymore so into the mists of git's history it goes.